Added another test to the test cases, after I noticed today that the tests were all passing even though I had a bug in the implementation.
I also added the functionality that the map returns an iterator to the item it just inserted and I added tests for this.

I also moved access to the map values to a function, so it hides the other
coolness I added, which is that the map will auto-delete an item if you pop
the last value off the stack for that key.