Added some actions to the application class interface.
I also tweaked the way you access the about window, now it's a static
slot, and you override a protected virtual function to customize the behavior.

I also moved the constructors to public, because over time there may be more
useful features to the application class that doesn't require the developer
to subclass it.