Made deep modifications to the Vector implementation.
The array pointer implementation has changed, and the vector's size on the
stack has changed.

Now you can easily make a vector with an interface by taking advantage of
the new template specialization features.  The syntax for templates is backwards compatible
with the old implementation.

You still have to implement all of the interfaces, but the infrastructure is
there now to make it easy.