Added a TreeView slot to resize columns
For some reason Qt doesn't already implement a slot to resize columns; you
have to manually call resizeSections() with the resize mode. So I
implemented a simple slot that just resizes all columns to their contents,
so you can connect the expanded() and collapsed() signals to it and
automatically resize the columns while still allowing the user to manually
resize them.