Expose the KDF and Salt from the Cryptor
It used to be that once you initialize a cryptor the Key Derivation
Function (KDF) was kept hidden along with its salt. I see no reason to
restrict the application developer in this way, so I exposed the KDF to
consumers of the Cryptor, and the KDF now exposes its salt. This makes
things more flexible so you can rely on the Cryptor to store and access
all this data, so you don't have to make your own copies for later
reference.