Fixed a bug in CheckCredentials
If there is a problem parsing the keyfile an exception could be thrown,
and it really should be regarded as a normal failure case (keyfile no
longer present = Creds Check Failed). Now all exceptions get trapped
and returned false.

I also added a new property on the cryptor that stores the credential
type, set with ChangeCredentials(). This information is useful to have
available.