Enable serial port functionality by default.
This is simply because the build broke in Qt 5.13, because qmake was
generating moc_.cc files for the serial port object even though it was
not defined (and this caused undefined reference errors).
I'm guessing Qt gave us another way to optionally exclude
the moc file generation, but I don't have time to figure that out, so
the hotfix is simply to include the serial port objects in the lib.